+++ NEWS.md
@@ -1,5 +1,7 @@
 # bstfun (development version)
 
+* Added `add_variable_grouping()` function that can group variable summaries together under a single header. (#79)
+
 * Adding new function `reinstall_prior_pkgs()` to assist in re-populating an R package library after a fresh install.
 
 # bstfun 0.3.5